direnv: Get rid of `nix-direnv`
`nix-direnv` is no longer 100% compatible with direnv's built in
use nix, which is kind of a bummer. It does not support passing
arguments such as `--system` and forwarding them to Nix itself.

One usecase is passing `--system x86_64-darwin` to `use nix` when
you're on macOS and want an emulated shell.

That's a shame, since it means that you can no longer share
`.envrc` files in repos that rely on this behavior and have them
be agnostic in the version of `use nix` that's being used.